+++ tests/api/test_statements_get.py
@@ -4,7 +4,6 @@ Tests for the GET statements endpoint of the Ralph API.
 import os
 import re
 from datetime import datetime, timedelta
-from unittest import mock
 from urllib.parse import quote_plus
 
 import pytest
@@ -14,6 +13,7 @@ from elasticsearch.helpers import bulk
 from fastapi.testclient import TestClient
 
 from ralph.api import app
+from ralph.defaults import Settings
 
 ES_TEST_HOSTS = os.environ.get("RALPH_ES_TEST_HOSTS", "http://localhost:9200").split(
     ","
@@ -269,12 +269,19 @@ def test_get_statements_until_timestamp(auth_credentials):
     assert response.json() == {"statements": [statements[0]]}
 
 
-@mock.patch("ralph.api.routers.statements.ES_MAX_SEARCH_HITS_COUNT", 2)
-def test_get_statements_with_pagination(auth_credentials):
+def test_get_statements_with_pagination(monkeypatch, auth_credentials):
     """
     When the first page does not contain all possible results, it includes
     a "more" property with a link to get the next page of results.
     """
+
+    def mock_settings():
+        """Mocks the get_settings function."""
+
+        return Settings(RUNSERVER_MAX_SEARCH_HITS_COUNT=2)
+
+    monkeypatch.setattr("ralph.api.routers.statements.get_settings", mock_settings)
+
     statements = [
         {
             "id": "5d345b99-517c-4b54-848e-45010904b177",