+++ tests/test_parsers.py
@@ -7,7 +7,7 @@ from io import StringIO
 
 import pytest
 
-from ralph.defaults import LOCALE_ENCODING
+from ralph.defaults import get_settings
 from ralph.parsers import GELFParser
 
 
@@ -31,7 +31,9 @@ def test_parsers_gelfparser_parse_raw_file(gelf_logger):
     gelf_logger.info('{"username": "foo"}')
     gelf_logger.info('{"username": ""}')
 
-    with open(gelf_logger.handlers[0].stream.name, encoding=LOCALE_ENCODING) as stream:
+    with open(
+        gelf_logger.handlers[0].stream.name, encoding=get_settings().LOCALE_ENCODING
+    ) as stream:
         events = list(parser.parse(stream))
 
     assert len(events) == 2
@@ -54,7 +56,9 @@ def test_parsers_gelfparser_parse_gzipped_file(fs, gelf_logger):
             shutil.copyfileobj(log_file, gzipped_log_file)
 
     parser = GELFParser()
-    with open(gelf_logger.handlers[0].stream.name, encoding=LOCALE_ENCODING) as stream:
+    with open(
+        gelf_logger.handlers[0].stream.name, encoding=get_settings().LOCALE_ENCODING
+    ) as stream:
         events = list(parser.parse(stream))
 
     assert len(events) == 2